backend/requirements.txt
python-dotenv
alembic
langgraph
langchain-openai
langgraph-checkpoint-postgres
psycopg[binary,pool]
fastapi
uvicorn[standard]
